Implemented Navigation with multi stack, interceptor ApiKeyInterceptor with token, UI compose functions for AppToolbar, AppNavigationBar and additional components. Implemented HomeScreen and Mocked functions for other screens.

+++ app/src/main/java/com/example/cryptos/data/network/entities/CoinListEntity.kt
@@ -1,7 +1,7 @@
 package com.example.cryptos.data.network.entities
 
 data class CoinListEntity(
-    val timestamp: Int? = null,
+    val timestamp: Long? = null,
     val data: List<CoinEntity>? = null,
 )
